requirements.txt
asgiref==3.11.0
django==6.0.1
django-cors-headers==4.9.0
django-unfold==0.77.1
djangorestframework==3.16.1
gunicorn==25.0.3
packaging==26.0
python-decouple==3.8
sqlparse==0.5.5
whitenoise==6.11.0
